+++ plone/app/dexterity/browser/modeleditor.py
@@ -3,8 +3,8 @@ from AccessControl import Unauthorized
 from lxml import etree
 from plone.app.dexterity import _
 from plone.supermodel.parser import SupermodelParseError
-from Products.Five import BrowserView
 from Products.CMFPlone.utils import safe_unicode
+from Products.Five import BrowserView
 from zope.component import queryMultiAdapter
 
 import json